@@ -1,8 +1,10 @@
-import ReactDOM from 'react-dom'
-import App from './App'
+import { createRoot } from 'react-dom/client'
+import { AppWithContexts } from './App'
 
 it('renders without crashing', () => {
   const div = document.createElement('div')
-  ReactDOM.render(<App />, div)
-  ReactDOM.unmountComponentAtNode(div)
+  const root = createRoot(div)
+
+  root.render(<AppWithContexts />)
+  root.unmount()
 })
